+++ app/Http/Livewire/Location/MediaForm.php
@@ -12,11 +12,16 @@ class MediaForm extends Component
 
     public Location $location;
     
-    public $mediaComponentNames = ['photos'];
-
+    public $mediaComponentNames = ['photos','contract', 'logo', 'icon', 'facade'];
     public $photos;
+    public $contract;
+    public $logo;
+    public $icon;
+    public $facade;
 
-    protected $rules = ['location.video' => 'nullable'];
+    protected $rules = [        
+        'location.video'    => 'nullable',        
+    ];
     
     public function save()
     {
@@ -24,9 +29,30 @@ class MediaForm extends Component
         
         $this->location->save();
         
-        $this->location->addFromMediaLibraryRequest($this->photos)->toMediaCollection('location-photos');
+        if ($this->photos) {
+            $this->location->addFromMediaLibraryRequest($this->photos)->toMediaCollection('location-photos');
+        }
         
-        session()->flash('success', 'Video updated successfully!');
+        if ($this->contract) {
+            $this->location->addFromMediaLibraryRequest($this->contract)
+            ->toMediaCollection('location-contracts');
+        }
+
+        if ($this->logo) {
+            $this->location->addFromMediaLibraryRequest($this->logo)->toMediaCollection('location-logo');
+            $this->location->logo = $this->location->getMedia('location-logo')->last()->getUrl();            
+        }
+
+        if ($this->icon) {
+            $this->location->addFromMediaLibraryRequest($this->icon)->toMediaCollection('location-icon');
+        }
+
+        if ($this->facade) {
+            $this->location->addFromMediaLibraryRequest($this->facade)->toMediaCollection('location-facade');
+            $this->location->facade = $this->location->getMedia('location-facade')->last()->getUrl();
+        }
+        
+        session()->flash('success', 'Media saved successfully!');
 
         $this->clearMedia();
     }