+++ resources/views/components/form/checkbox.blade.php
@@ -4,7 +4,7 @@
             class="focus:ring-indigo-500 h-4 w-4 text-indigo-600 border-gray-300 rounded">
     </div>
     <div class="ml-3 text-sm">
-        <label for="candidates" class="font-medium text-gray-500 capitalize">{{ $label ?? $name }}</label>
+        <label for="candidates" class="font-medium text-gray-700 capitalize">{{ $label ?? $name }}</label>
         @if ($description)
         <p class="text-gray-500">{{ $description }}</p>
         @endif
@@ -12,7 +12,7 @@
 </div>
 
 
-{{-- 
+{{--
 <div class="flex items-center">
     <div class="flex items-center h-5">
         <input wire:model="thursday" type="checkbox" class="form-checkbox df-form-checkbox">