+++ resources/views/location/show/sidebar.blade.php
@@ -4,11 +4,14 @@
             <x-location.description-card :location="$location" />
         </div>
         <div class="pt-2 space-y-6">
+            @if ($location->contract)
             <div>
                 <h3 class="font-medium text-gray-900">Contract</h3>
                 <a href="{{ asset($location->contract) }}" target="_blank"
                     class="block bg-indigo-600 text-white text-center py-2 mt-2 rounded-lg">Download</a>
             </div>
+            @endif
+
             @if ($location->comments)
             <div>
                 <h3 class="font-medium text-gray-900">Comments</h3>