+++ resources/views/pages/components.blade.php
@@ -1,0 +1,34 @@
+<h1>List of components</h1>
+<table width="100%">
+    <thead>
+        <th>Laravel</th>
+        <th>Livewire</th>
+    </thead>
+    <tbody>
+        <tr>
+            <td>Instructors list</td>
+            <td>Photo Gallery</td>
+        </tr>
+        <tr>
+            <td>Select user</td>
+            <td></td>
+        </tr>
+        <tr>
+            <td></td>
+            <td></td>
+        </tr>
+        <tr>
+            <td></td>
+            <td></td>
+        </tr>
+        <tr>
+            <td></td>
+            <td></td>
+        </tr>
+        <tr>
+            <td></td>
+            <td></td>
+        </tr>
+
+    </tbody>
+</table>